// src/api/models/FreightDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PK.PHCalc.OutsourcingDummy
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    Bounds,
    BoundsFromJSON,
    BoundsFromJSONTyped,
    BoundsToJSON,
} from './Bounds';

/**
 * 
 * @export
 * @interface FreightDTO
 */
export interface FreightDTO {
    /**
     * 
     * @type {number}
     * @memberof FreightDTO
     */
    price: number;
    /**
     * 
     * @type {Bounds}
     * @memberof FreightDTO
     */
    totalAreaSquareMeterBounds: Bounds;
    /**
     * 
     * @type {number}
     * @memberof FreightDTO
     */
    qualityCategory: FreightDTOQualityCategoryEnum;
}

/**
* @export
* @enum {string}
*/
export enum FreightDTOQualityCategoryEnum {
    NUMBER_0 = 0,
    NUMBER_1 = 1,
    NUMBER_2 = 2,
    NUMBER_3 = 3
}

export function FreightDTOFromJSON(json: any): FreightDTO {
    return FreightDTOFromJSONTyped(json, false);
}

export function FreightDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): FreightDTO {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'price': json['Price'],
        'totalAreaSquareMeterBounds': BoundsFromJSON(json['TotalAreaSquareMeterBounds']),
        'qualityCategory': json['QualityCategory'],
    };
}

export function FreightDTOToJSON(value?: FreightDTO | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'Price': value.price,
        'TotalAreaSquareMeterBounds': BoundsToJSON(value.totalAreaSquareMeterBounds),
        'QualityCategory': value.qualityCategory,
    };
}